refactor: reorganize button and badge variants, move form context to utils
