+++ src/components/ItemRow.tsx
@@ -9,7 +9,7 @@ export interface ItemRowProps {
   onOpen?: () => void;
 }
 
-const ItemRow: React.FC<ItemRowProps> = ({ id: _id, name, description, quantity, thumbnailUrl, onOpen }) => (
+const ItemRow: React.FC<ItemRowProps> = ({ name, description, quantity, thumbnailUrl, onOpen }) => (
   <div className="flex items-center p-2 border-b cursor-pointer" onClick={onOpen} aria-label={`Open item ${name}`}>
     {thumbnailUrl && <img src={thumbnailUrl} alt={name} className="w-8 h-8 rounded mr-2" />}
     <div className="flex-1">
