+++ src/components/ui/form-utils.ts
@@ -1,0 +1,51 @@
+import * as React from "react";
+import {
+    type FieldPath,
+    type FieldValues,
+    useFormContext,
+    useFormState,
+} from "react-hook-form";
+
+// Shared contexts extracted from form.tsx so that form.tsx only exports components (satisfies react-refresh/only-export-components)
+
+export type FormFieldContextValue<
+    TFieldValues extends FieldValues = FieldValues,
+    TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
+> = {
+    name: TName;
+};
+
+export const FormFieldContext = React.createContext<FormFieldContextValue>(
+    {} as FormFieldContextValue,
+);
+
+export type FormItemContextValue = {
+    id: string;
+};
+
+export const FormItemContext = React.createContext<FormItemContextValue>(
+    {} as FormItemContextValue,
+);
+
+export const useFormField = () => {
+    const fieldContext = React.useContext(FormFieldContext);
+    const itemContext = React.useContext(FormItemContext);
+    const { getFieldState } = useFormContext();
+    const formState = useFormState({ name: fieldContext.name });
+    const fieldState = getFieldState(fieldContext.name, formState);
+
+    if (!fieldContext) {
+        throw new Error("useFormField should be used within <FormField>");
+    }
+
+    const { id } = itemContext;
+
+    return {
+        id,
+        name: fieldContext.name,
+        formItemId: `${id}-form-item`,
+        formDescriptionId: `${id}-form-item-description`,
+        formMessageId: `${id}-form-item-message`,
+        ...fieldState,
+    };
+};
